+++ one-third-hal/core/uart/usart1.c
@@ -68,8 +68,8 @@ static void Usart1RingConfig(uint8_t* data, uint16_t len) {
 
 // ----------------------------------------------------------------------------
 static void Usart1DmaConfig(uint8_t* buffer, uint16_t len) {
-    ( void )buffer;
-    ( void )len;
+    (void)buffer;
+    (void)len;
     // if we use DMA for receiving, we don't want UART_IT_RXNE
     __HAL_UART_DISABLE_IT(&(usart1.huart), UART_IT_RXNE);
     __HAL_RCC_DMA2_CLK_ENABLE();
@@ -94,7 +94,7 @@ static void Usart1DmaConfig(uint8_t* buffer, uint16_t len) {
     // only use verified code
 #if defined(STM32F407xx)
     HAL_DMA_Start(&hdma_usart1_rx, (uint32_t) & (usart1.huart.Instance->DR),
-                  ( uint32_t )buffer, len);
+                  (uint32_t)buffer, len);
 #elif defined(STM32F767xx)
     // HAL_DMA_Start(&hdma_usart2_rx, (uint32_t) & (usart2.huart.Instance->RDR),
     //               ( uint32_t )buffer, len);
@@ -131,20 +131,20 @@ static void Usart1IdleIrqCallback(void) {
         static uint16_t remaining[2] = { 0, 0 };
         uint16_t receive_count = 0;
         remaining[0] = remaining[1];
-        remaining[1] = ( uint16_t )(__HAL_DMA_GET_COUNTER(usart1.huart.hdmarx));
+        remaining[1] = (uint16_t)(__HAL_DMA_GET_COUNTER(usart1.huart.hdmarx));
         static bool dma_first_packet = true;
         if (dma_first_packet) {
-            receive_count = usart1.rb.state.capacity - remaining[1];
+            receive_count = (uint16_t)(usart1.rb.state.capacity - remaining[1]);
             dma_first_packet = false;
             op.ringbuffer.pushN(&(usart1.rb), usart1.rb.data, receive_count);
         }
         else {
             if (remaining[0] > remaining[1]) {
-                receive_count = remaining[0] - remaining[1];
+                receive_count = (uint16_t)(remaining[0] - remaining[1]);
             }
             else {
-                receive_count = ( uint16_t )(usart1.rb.state.capacity
-                                             + remaining[0] - remaining[1]);
+                receive_count = (uint16_t)(usart1.rb.state.capacity
+                                           + remaining[0] - remaining[1]);
             }
             // DMA has already pushed data into the ringbuffer
             op.ringbuffer.added(&(usart1.rb), receive_count);
